third_party/WebKit/LayoutTests/svg/filters/feConvolveMatrix-order-error-reporting-expected.txt

CONSOLE ERROR: line 7: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0")
CONSOLE ERROR: line 8: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0.49999999")
CONSOLE ERROR: line 9: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0.5")
CONSOLE ERROR: line 10: Error: <feConvolveMatrix> attribute order: A negative value is not valid. ("-1")
CONSOLE ERROR: line 11: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0")
CONSOLE ERROR: line 12: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0.49999999")
CONSOLE ERROR: line 13: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0.5")
CONSOLE ERROR: line 14: Error: <feConvolveMatrix> attribute order: A negative value is not valid. ("1 -1")
CONSOLE ERROR: line 15: Error: <feConvolveMatrix> attribute order: Expected integer, "1 one".
CONSOLE ERROR: line 16: Error: <feConvolveMatrix> attribute order: Expected integer, "one two".
